requirements.txt
PyQt6
python-chess
